# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fabiq.SmartFactory.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Downtime/recent": {
      "get": {
        "tags": [
          "Downtime"
        ],
        "parameters": [
          {
            "name": "lineId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DowntimeEventDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DowntimeEventDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DowntimeEventDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Downtime/summary": {
      "get": {
        "tags": [
          "Downtime"
        ],
        "parameters": [
          {
            "name": "lineId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DowntimeSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DowntimeSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DowntimeSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Downtime": {
      "post": {
        "tags": [
          "Downtime"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDowntimeEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDowntimeEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDowntimeEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Oee/line/{lineId}": {
      "get": {
        "tags": [
          "Oee"
        ],
        "parameters": [
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plannedProductionMinutes",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 120
            }
          },
          {
            "name": "idealCycleTimeSeconds",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Oee/work-order/{workOrderId}": {
      "get": {
        "tags": [
          "Oee"
        ],
        "parameters": [
          {
            "name": "workOrderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plannedProductionMinutes",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 120
            }
          },
          {
            "name": "idealCycleTimeSeconds",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OeeDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateDowntimeEventRequest": {
        "type": "object",
        "properties": {
          "machineId": {
            "type": "string",
            "nullable": true
          },
          "workOrderId": {
            "type": "string",
            "nullable": true
          },
          "reasonCode": {
            "type": "string",
            "nullable": true
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DowntimeEventDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "nullable": true
          },
          "machineId": {
            "type": "string",
            "nullable": true
          },
          "lineId": {
            "type": "string",
            "nullable": true
          },
          "workOrderId": {
            "type": "string",
            "nullable": true
          },
          "reasonCode": {
            "type": "string",
            "nullable": true
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DowntimeSummaryDto": {
        "type": "object",
        "properties": {
          "lineId": {
            "type": "string",
            "nullable": true
          },
          "reasonCode": {
            "type": "string",
            "nullable": true
          },
          "eventCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "affectedMachineCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastOccurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "OeeDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "nullable": true
          },
          "plannedProductionMinutes": {
            "type": "number",
            "format": "double"
          },
          "runningMinutes": {
            "type": "number",
            "format": "double"
          },
          "idealCycleTimeSeconds": {
            "type": "number",
            "format": "double"
          },
          "goodCount": {
            "type": "integer",
            "format": "int32"
          },
          "scrapCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "availability": {
            "type": "number",
            "format": "double"
          },
          "performance": {
            "type": "number",
            "format": "double"
          },
          "quality": {
            "type": "number",
            "format": "double"
          },
          "oee": {
            "type": "number",
            "format": "double"
          },
          "calculatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}